Check video url exist before save, refacto into a VideoService
